+++ pkg/awsglue/glue.go
@@ -22,6 +22,10 @@ import (
 	"time"
 
 	"github.com/aws/aws-sdk-go/aws"
+	"github.com/aws/aws-sdk-go/aws/awserr"
+	"github.com/aws/aws-sdk-go/service/glue"
+	"github.com/aws/aws-sdk-go/service/glue/glueiface"
+	"github.com/pkg/errors"
 )
 
 const (
@@ -40,80 +44,74 @@ const (
 	GlueTimestampType = "timestamp" // type in Glue tables for timestamps that we will re-map Go times
 )
 
-// Use this to tag the time partitioning used in a Glue table
-type GlueTableTimebin int
-
-const (
-	GlueTableMonthly GlueTableTimebin = iota + 1
-	GlueTableDaily
-	GlueTableHourly
-)
-
-func (tb GlueTableTimebin) Validate() (err error) {
-	switch tb {
-	case GlueTableHourly, GlueTableDaily, GlueTableMonthly:
-		return
-	default:
-		err = fmt.Errorf("unknown Glue table time bin: %d", tb)
-	}
-	return
+type PartitionKey struct {
+	Name string
+	Type string
 }
 
-// return the next time interval
-func (tb GlueTableTimebin) Next(t time.Time) (next time.Time) {
-	switch tb {
-	case GlueTableHourly:
-		return t.Add(time.Hour).Truncate(time.Hour)
-	case GlueTableDaily:
-		return t.Add(time.Hour * 24).Truncate(time.Hour * 24)
-	case GlueTableMonthly:
-		// loop a day at a time until the month changes
-		currentMonth := t.Month()
-		for next = t.Add(time.Hour * 24).Truncate(time.Hour * 24); next.Month() == currentMonth; next = next.Add(time.Hour * 24) {
-		}
-		return next
-	default:
-		panic(fmt.Sprintf("unknown Glue table time bin: %d", tb))
-	}
-}
-
-// Meta data about Glue table over parser data written to S3
+// Meta data about GlueTableMetadata table over parser data written to S3
 // NOTE: this struct has all accessor behind functions to allow a lazy evaluation
 //       so the cost of creating the schema is only when actually needing this information.
-type GlueMetadata struct {
-	databaseName string
-	tableName    string
-	description  string
-	s3Prefix     string           // where we expect to find data relative to the bucket (excluding time partitions)
-	timebin      GlueTableTimebin // at what time resolution is this table partitioned
-	timeUnpadded bool             // if true, do not zero pad partition time values
-	eventStruct  interface{}      // object used to infer columns
+type GlueTableMetadata struct {
+	databaseName  string
+	tableName     string
+	description   string
+	logType string
+	s3TablePrefix string           // All data for this table are stored in this S3 prefix
+	timebin       GlueTableTimebin // at what time resolution is this table partitioned
+	timeUnpadded  bool             // if true, do not zero pad partition time values
+	eventStruct interface{}
 }
 
-func (gm *GlueMetadata) DatabaseName() string {
+func (gm *GlueTableMetadata) DatabaseName() string {
 	return gm.databaseName
 }
 
-func (gm *GlueMetadata) TableName() string {
+func (gm *GlueTableMetadata) TableName() string {
 	return gm.tableName
 }
 
-func (gm *GlueMetadata) Description() string {
+func (gm *GlueTableMetadata) Description() string {
 	return gm.description
 }
 
-func (gm *GlueMetadata) S3Prefix() string {
-	return gm.s3Prefix
+func (gm *GlueTableMetadata) S3Prefix() string {
+	return gm.s3TablePrefix
 }
 
-func (gm *GlueMetadata) Timebin() GlueTableTimebin {
+func (gm *GlueTableMetadata) Timebin() GlueTableTimebin {
 	return gm.timebin
 }
 
+func (gm *GlueTableMetadata) Timebin() GlueTableTimebin {
+	return gm.timebin
+}
+
+func (gm *GlueTableMetadata) EventStruct() interface{} {
+	return gm.eventStruct
+}
+
+func (gm *GlueTableMetadata) PartitionKeys() (partitions []PartitionKey) {
+	partitions = []PartitionKey{
+		{Name: "year", Type: "int"},
+	}
+
+	if gm.Timebin() >= GlueTableMonthly {
+		partitions = append(partitions, PartitionKey{Name: "month", Type: "int"})
+	}
+	if gm.Timebin() >= GlueTableDaily {
+		partitions = append(partitions, PartitionKey{Name: "day", Type: "int"})
+	}
+	if gm.Timebin() >= GlueTableHourly {
+		partitions = append(partitions, PartitionKey{Name: "hour", Type: "int"})
+	}
+	return partitions
+}
+
 // Based on Timebin(), return an S3 prefix for objects
-func (gm *GlueMetadata) PartitionPrefix(t time.Time) (prefix string) {
-	partitionValues := gm.PartitionValues(t)
-	prefix = gm.s3Prefix
+func (gm *GlueTableMetadata) PartitionPrefix(t time.Time) (prefix string) {
+	partitionValues := gm.partitionValues(t)
+	prefix = gm.s3TablePrefix
 	switch gm.timebin {
 	case GlueTableHourly:
 		prefix += fmt.Sprintf("year=%s/month=%s/day=%s/hour=%s/",
@@ -128,30 +126,8 @@ func (gm *GlueMetadata) PartitionPrefix(t time.Time) (prefix string) {
 	return
 }
 
-type Partition struct {
-	Name string
-	Type string
-}
-
-func (gm *GlueMetadata) PartitionKeys() (partitions []Partition) {
-	partitions = []Partition{
-		{Name: "year", Type: "int"},
-	}
-
-	if gm.Timebin() >= GlueTableMonthly {
-		partitions = append(partitions, Partition{Name: "month", Type: "int"})
-	}
-	if gm.Timebin() >= GlueTableDaily {
-		partitions = append(partitions, Partition{Name: "day", Type: "int"})
-	}
-	if gm.Timebin() >= GlueTableHourly {
-		partitions = append(partitions, Partition{Name: "hour", Type: "int"})
-	}
-	return partitions
-}
-
-// Based on Timebin(), return an []*string values (used for Glue APIs)
-func (gm *GlueMetadata) PartitionValues(t time.Time) (values []*string) {
+// Based on Timebin(), return an []*string values (used for GlueTableMetadata APIs)
+func (gm *GlueTableMetadata) partitionValues(t time.Time) (values []*string) {
 	var intFormat string
 
 	if gm.timeUnpadded {
@@ -174,43 +150,99 @@ func (gm *GlueMetadata) PartitionValues(t time.Time) (values []*string) {
 	return
 }
 
-func (gm *GlueMetadata) EventStruct() interface{} {
-	return gm.eventStruct
-}
-
-// Clone returns a copy of table with s3Prefix and database changed
-func (gm *GlueMetadata) Clone(s3Prefix, databaseName string) *GlueMetadata {
-	clone := *gm // copy
-	clone.s3Prefix = standardizeS3Prefix(s3Prefix, gm.TableName())
-	clone.databaseName = databaseName
-	return &clone
-}
-
-func NewGlueMetadata(s3Prefix, databaseName, tableName, description string, timebin GlueTableTimebin,
-	timeUnpadded bool, eventStruct interface{}) (gm *GlueMetadata, err error) {
-
-	err = timebin.Validate()
-	if err != nil {
-		return
+func NewLogTableMetadata(logType, description string) *GlueTableMetadata {
+	return &GlueTableMetadata{
+		databaseName:  LogProcessingDatabaseName,
+		tableName:     standardizeTableName(logType),
+		description:   description,
+		s3TablePrefix: standardizeS3Prefix(LogS3Prefix, logType),
+		timebin:       GlueTableHourly,
+		timeUnpadded:  false,
 	}
+}
 
+func NewRuleTableMetadata(logType, description string) *GlueTableMetadata {
+	return &GlueTableMetadata{
+		databaseName:  RuleMatchDatabaseName,
+		tableName:     standardizeTableName(logType),
+		description:   description,
+		s3TablePrefix: standardizeS3Prefix(RuleMatchS3Prefix, logType),
+		timebin:       GlueTableHourly,
+		timeUnpadded:  false,
+	}
+}
+
+func standardizeTableName(logType string) string {
 	// clean table name to make sql friendly
-	tableName = strings.Replace(tableName, ".", "_", -1) // no '.'
-	tableName = strings.ToLower(tableName)
-
-	gm = &GlueMetadata{
-		databaseName: databaseName,
-		tableName:    tableName,
-		description:  description,
-		s3Prefix:     standardizeS3Prefix(s3Prefix, tableName),
-		timebin:      timebin,
-		timeUnpadded: timeUnpadded,
-		eventStruct:  eventStruct,
-	}
-
-	return
+	tableName := strings.Replace(logType, ".", "_", -1) // no '.'
+	return strings.ToLower(tableName)
 }
 
-func standardizeS3Prefix(s3Prefix, tableName string) string {
-	return s3Prefix + "/" + tableName + "/" // ensure last char is '/'
+func standardizeS3Prefix(s3Prefix, logType string) string {
+	return s3Prefix + "/" + standardizeTableName(logType) + "/" // ensure last char is '/'
+}
+
+// CreateJSONPartition creates a new JSON partition in a GlueTableMetadata table. If the partition already exists, no partition is added.
+func (gm *GlueTableMetadata) CreateJSONPartition(client glueiface.GlueAPI, t time.Time) error {
+	partitionPrefix := "s3://" +  gm.PartitionPrefix(t)
+
+	partitionInput := &glue.PartitionInput{
+		Values:            gm.partitionValues(t),
+		StorageDescriptor: getJSONPartitionDescriptor(partitionPrefix),
+	}
+	input := &glue.CreatePartitionInput{
+		DatabaseName:   aws.String(gm.databaseName),
+		TableName:      aws.String(gm.tableName),
+		PartitionInput: partitionInput,
+	}
+	_, err := client.CreatePartition(input)
+	if err != nil {
+		if awsErr, ok := err.(awserr.Error); ok {
+			if awsErr.Code() == "AlreadyExistsException" {
+				return nil
+			}
+		}
+		return errors.Wrap(err, "failed to create new JSON partition")
+	}
+	return err
+}
+
+
+// SyncPartition deletes and re-creates a partition using the latest table schema. Used when schemas change.
+func (gm *GlueTableMetadata) SyncPartition(client glueiface.GlueAPI, t time.Time) error {
+	_, err := gm.deletePartition(client, t)
+	if err != nil {
+		if awsErr, ok := err.(awserr.Error); !ok || awsErr.Code() != "EntityNotFoundException" {
+			return errors.Wrapf(err, "delete partition for %s.%s at %v failed", gm.DatabaseName(), gm.TableName(), t)
+		}
+	}
+	err = gm.CreateJSONPartition(client, t)
+	if err != nil {
+		return errors.Wrapf(err, "create partition for %s.%s at %v failed", gm.DatabaseName(), gm.TableName(), t)
+	}
+	return nil
+}
+
+func (gm *GlueTableMetadata) deletePartition(client glueiface.GlueAPI, t time.Time) (output *glue.DeletePartitionOutput, err error) {
+	input := &glue.DeletePartitionInput{
+		DatabaseName:    aws.String(gm.databaseName),
+		TableName:       aws.String(gm.tableName),
+		PartitionValues: gm.partitionValues(t),
+	}
+	return client.DeletePartition(input)
+}
+
+func getJSONPartitionDescriptor(s3Path string) *glue.StorageDescriptor {
+	return &glue.StorageDescriptor{
+		InputFormat:  aws.String("org.apache.hadoop.mapred.TextInputFormat"),
+		OutputFormat: aws.String("org.apache.hadoop.hive.ql.io.HiveIgnoreKeyTextOutputFormat"),
+		SerdeInfo: &glue.SerDeInfo{
+			SerializationLibrary: aws.String("org.openx.data.jsonserde.JsonSerDe"),
+			Parameters: map[string]*string{
+				"serialization.format": aws.String("1"),
+				"case.insensitive":     aws.String("TRUE"), // treat as lower case
+			},
+		},
+		Location: aws.String(s3Path),
+	}
 }
